Produce Standard MIDI Format (SMF) compliant MIDI files
Migrate build.sh from 0.9.7 to permit building on OS X.

Modify SMF.h, SMFEvent.h, smf_event.cpp, & smf.cpp to generate the MIDI
Tempo track as the first track and populate it with the Copyright Notice
,Time Signature & Set Tempo meta events. The Track Name meta event is
moved to the tempo map.

The name in the Copyright Notice meta event will be extracted from the
Song properites as filled in at the Project->Show Info dialog, which
will be preceded with the copyright (C) symbol and followed with the
current year as specified in the SMF standard. This will be the first
event in the file, also as specified in the SMF standard.

The Track Name meta event will now be extracted from the Song properites
as filled in at the Project->Show Info dialog and will be located in the
tempo map as per the SMF standard.

The Time Signature meta event will be created assuming 4/4 time.

The note on & note off events will start in the second track in the
file.